+++ app/src/main/assets/api.js
@@ -1,8 +1,8 @@
-const serverURL = window.location.href
-// const serverURL = "http://192.168.1.102:8080"
+// const serverURL = window.location.href
+const serverURL = "http://192.168.1.102:8080/"
 const headers = {
-  'Accept': 'application/json',
-  'Content-Type': 'application/json'
+  Accept: "application/json",
+  "Content-Type": "application/json",
 }
 
 // ============================================
@@ -13,10 +13,10 @@ function _request_post(path, body) {
   const options = {
     method: "post",
     headers,
-    body: JSON.stringify(body)
+    body: JSON.stringify(body),
   }
 
-  return fetch(`${serverURL}/${path}`, options)
+  return fetch(`${serverURL}${path}`, options)
 }
 
 function _request_get(path) {
@@ -25,30 +25,33 @@ function _request_get(path) {
     headers,
   }
 
-  return fetch(`${serverURL}/${path}`, options)
+  return fetch(`${serverURL}${path}`, options)
 }
 
 // ============================================
 // 									REQUESTS
 // ============================================
 
+// ================= PLAYER ===================
 
-const playCurrentStation = () => _request_post('current/play')
-  .then(response => response.json())
+const playCurrentStation = () => _request_post("current/play").then((response) => response.json())
 
-const stopCurrentStation = () => _request_post('current/stop')
-  .then(response => response.json())
+const stopCurrentStation = () => _request_post("current/stop").then((response) => response.json())
 
-const loadCurrentStation = () => 			_request_get('station/current')
-  .then(response => response.json())
+const loadCurrentStation = () => _request_get("station/current").then((response) => response.json())
 
-const updateRadioVolume = (Percentage) =>	_request_post('volume/level', { Percentage })
+const updateRadioVolume = (Percentage) => _request_post("volume/level", { Percentage })
 
+// ============= SELECTED STATION =============
 
-const playStationByUUID = (UUID) => _request_post('station/play', { UUID })
+const playStationByUUID = (UUID) => _request_post("station/play", { UUID }).then((response) => response.json())
 
-const deleteStationByUUID = (UUID) =>	_request_post('station/delete', { UUID })
-  .then(response => response.json())
+const deleteStationByUUID = (UUID) => _request_post("station/delete", { UUID }).then((response) => response.json())
 
-const saveNewStation = (URL, Title) => _request_post('station/save', { URL, Title })
-  .then(response => response.json())
+const saveNewStation = (URL, Title) => _request_post("station/save", { URL, Title }).then((response) => response.json())
+
+// ============= IMPORT / EXPORT ==============
+
+const exportStations = () => _request_get("list/export").then((response) => response.json())
+
+const importStations = (data) => _request_post("list/import", data).then((response) => response.json())