+++ app/src/main/java/com/dejssa/radioserver/Service/Server.java
@@ -11,11 +11,15 @@ import com.dejssa.radioserver.storage.requests.StationRequest;
 import com.dejssa.radioserver.storage.requests.VolumeLevelRequest;
 import com.dejssa.radioserver.storage.responses.StatusResponse;
 import com.google.gson.Gson;
+import com.google.gson.reflect.TypeToken;
 
 import java.io.IOException;
 import java.lang.reflect.Type;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.HashMap;
+import java.util.List;
+import java.util.UUID;
 import java.util.stream.IntStream;
 
 import fi.iki.elonen.NanoHTTPD;
@@ -36,55 +40,66 @@ public class Server extends NanoHTTPD{
 
     @Override
     public Response serve(IHTTPSession session) {
-        if (session.getUri().contains("current/play")) {
+        // Play current station.
+        if (compare(session, "/current/play")) {
             this.playAudio();
-
             return newFixedLengthResponse(new Gson().toJson(this.prepareRadioStatus()));
         }
 
-        if (session.getUri().contains("current/stop")) {
+        // Stop playing current station.
+        if (compare(session, "/current/stop")) {
             this.stopAudio();
-
             return newFixedLengthResponse(new Gson().toJson(this.prepareRadioStatus()));
         }
 
-        if (session.getUri().contains("volume/level")) {
+        // Change volume of the device.
+        if (compare(session, "/volume/level")) {
             VolumeLevelRequest request = this.parseRequest(session, VolumeLevelRequest.class);
-
             this.setVolumeTo(request.Percentage);
-
             return newFixedLengthResponse("");
         }
 
-        if (session.getUri().contains("station/current"))  {
+        // Current station info was requested.
+        if (compare(session, "/station/current")) {
             return newFixedLengthResponse(new Gson().toJson(this.prepareRadioStatus()));
         }
 
-        if (session.getUri().contains("station/save")) {
+        // Save a new station.
+        if (compare(session, "/station/save")) {
             StationRequest request = this.parseRequest(session, StationRequest.class);
-
             this.Stations.add(request.toDomain());
-
             return newFixedLengthResponse(new Gson().toJson(this.Stations));
         }
 
-        if (session.getUri().contains("station/delete")) {
+        // Remove selected station.
+        if (compare(session, "/station/delete")) {
             StationUUIDRequest request = this.parseRequest(session, StationUUIDRequest.class);
-
             deleteSelectedStation(request);
-
             return newFixedLengthResponse(new Gson().toJson(this.Stations));
         }
 
-        if (session.getUri().contains("station/play")) {
+        // Play selected station.
+        if (compare(session, "/station/play")) {
             StationUUIDRequest request = this.parseRequest(session, StationUUIDRequest.class);
-
             playStationByUUID(request);
-
             return newFixedLengthResponse(new Gson().toJson(this.prepareRadioStatus()));
+        }
+
+        // Export the list of stations.
+        if (compare(session, "/list/export")) {
+            return newFixedLengthResponse(new Gson().toJson(this.Stations));
+        }
+
+        // Import a new list of stations.
+        if (compare(session, "/list/import")) {
+            return serverImportStations(session);
         }
 
         return this.webAppResponse.serve(session);
+    }
+
+    private boolean compare(IHTTPSession session, String path) {
+        return session.getUri().compareToIgnoreCase(path) == 0;
     }
 
     @Override
@@ -97,6 +112,21 @@ public class Server extends NanoHTTPD{
     public void stop() {
         stopAudio();
         super.stop();
+    }
+
+    private Response serverImportStations(IHTTPSession session) {
+//        ArrayList<StationRequest> request = this.parseRequest(session, ArrayList.class);
+        List<StationRequest> request = this.parseArrayRequest(session, StationRequest[].class);
+
+        ArrayList<StationInfo> newStations = new ArrayList<>();
+
+        for (int i = 0; i < request.size(); i++) {
+            newStations.add(request.get(i).toDomain());
+        }
+
+        this.Stations = newStations;
+
+        return newFixedLengthResponse(new Gson().toJson(this.Stations));
     }
 
     private void playAudio() {
@@ -185,5 +215,16 @@ public class Server extends NanoHTTPD{
 
 
        return new Gson().fromJson(request.get("postData"), (Type) classOfT);
+    }
+
+    private <T> List<T> parseArrayRequest(NanoHTTPD.IHTTPSession session, Class<T[]> classOfT) {
+        HashMap<String, String> request = new HashMap<>();
+        try {
+            session.parseBody(request);
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+
+        return Arrays.asList(new Gson().fromJson(request.get("postData"), classOfT));
     }
 }